changed implementation from double to float

+++ MCO2-LBYARCH/main.c
@@ -1,16 +1,16 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-extern void saxpy_x86_64(int, double*, double*, double*);
+extern void saxpy_x86_64(int, float*, float*, float*);
 
 int main() {
 	const int N = 1 << 20; // 2^n change the second operand depending on N
 	const size_t ARRAY_SIZE = N * sizeof(double); 
 
-	double *vectorX, *vectorY, *vectorZ;
-	vectorX = (double*)malloc(ARRAY_SIZE);
-	vectorY = (double*)malloc(ARRAY_SIZE);
-	vectorZ = (double*)malloc(ARRAY_SIZE);
+	float *vectorX, *vectorY, *vectorZ;
+	vectorX = (float*)malloc(ARRAY_SIZE);
+	vectorY = (float*)malloc(ARRAY_SIZE);
+	vectorZ = (float*)malloc(ARRAY_SIZE);
 
 	if (vectorX == NULL || vectorY == NULL || vectorZ == NULL) {
 		printf("Memory Allocation Failed");
@@ -18,7 +18,7 @@ int main() {
 	}
 
 	int counter;
-	double i, j;
+	float i, j;
 	for (counter = 0, i = 1.0, j = 11.0; counter < N; counter++, i++, j++) {
 		vectorX[counter] = i;
 		vectorY[counter] = j;
@@ -28,7 +28,7 @@ int main() {
 	saxpy_x86_64(N, vectorX, vectorY, vectorZ);
 
 	for (counter = 0; counter < 10; counter++) {
-		printf("Output %d: %lf\n", counter + 1, vectorZ[counter]);
+		printf("Output %d: %f\n", counter + 1, vectorZ[counter]);
 	}
 
 	free(vectorX);
